Refactor JLink SWD handling code
